Move neet.sensitivity to neet.boolean.sensitivity
